Add `::Pair` dispatch to `escapeuri`
The docstring for `URIs.URI` claims that passing a single query argument
as a `Pair` is valid. However, since there is no method specifically for
pairs the fallback method is used. The fallback method assumes the
argument iterates key-value pairs, but since `Pair` iterates the two
components this will simply encode the first characters of the key-value
pair instead.